+++ source/hwPortUtils.py
@@ -9,65 +9,31 @@ import itertools
 import math
 import typing
 import winreg
-from ctypes.wintypes import BOOL, DWORD, HWND, PDWORD, ULONG, USHORT, WCHAR
+from ctypes.wintypes import BOOL, DWORD, ULONG, USHORT, WCHAR
 
 from comtypes import GUID
 
+from winBindings.setupapi import SetupDiDestroyDeviceInfoList
 import config
 import hidpi
+from winBindings.setupapi import HDEVINFO
+from winBindings.setupapi import SetupDiGetClassDevs
+from winBindings.setupapi import SetupDiGetDeviceProperty
+from winBindings.setupapi import DEVPROPKEY
+from winBindings.setupapi import SP_DEVINFO_DATA
+from winBindings.setupapi import SetupDiEnumDeviceInterfaces
+from winBindings.setupapi import SP_DEVICE_INTERFACE_DATA
+from winBindings.setupapi import SetupDiGetDeviceInterfaceDetail
+from winBindings.setupapi import SetupDiGetDeviceRegistryProperty
 import winKernel
 from logHandler import log
 from winAPI.constants import SystemErrorCodes
 from winKernel import SYSTEMTIME
 
-
 def ValidHandle(value):
 	if value == 0:
 		raise ctypes.WinError()
 	return value
-
-
-HDEVINFO = ctypes.c_void_p
-
-
-class SP_DEVINFO_DATA(ctypes.Structure):
-	_fields_ = (
-		("cbSize", DWORD),
-		("ClassGuid", GUID),
-		("DevInst", DWORD),
-		("Reserved", ctypes.POINTER(ULONG)),
-	)
-
-	def __str__(self):
-		return f"ClassGuid:{self.ClassGuid} DevInst:{self.DevInst}"
-
-
-PSP_DEVINFO_DATA = ctypes.POINTER(SP_DEVINFO_DATA)
-
-
-class SP_DEVICE_INTERFACE_DATA(ctypes.Structure):
-	_fields_ = (
-		("cbSize", DWORD),
-		("InterfaceClassGuid", GUID),
-		("Flags", DWORD),
-		("Reserved", ctypes.POINTER(ULONG)),
-	)
-
-	def __str__(self):
-		return f"InterfaceClassGuid:{self.InterfaceClassGuid} Flags:{self.Flags}"
-
-
-PSP_DEVICE_INTERFACE_DATA = ctypes.POINTER(SP_DEVICE_INTERFACE_DATA)
-
-PSP_DEVICE_INTERFACE_DETAIL_DATA = ctypes.c_void_p
-
-
-class DEVPROPKEY(ctypes.Structure):
-	_fields_ = (
-		("DEVPROPGUID", GUID),
-		("DEVPROPID", ULONG),
-	)
-
 
 class dummy(ctypes.Structure):
 	_fields_ = (("d1", DWORD), ("d2", WCHAR))
@@ -77,67 +43,6 @@ class dummy(ctypes.Structure):
 
 SIZEOF_SP_DEVICE_INTERFACE_DETAIL_DATA_W = ctypes.sizeof(dummy)
 
-SetupDiDestroyDeviceInfoList = ctypes.windll.setupapi.SetupDiDestroyDeviceInfoList
-SetupDiDestroyDeviceInfoList.argtypes = (HDEVINFO,)
-SetupDiDestroyDeviceInfoList.restype = BOOL
-
-SetupDiGetClassDevs = ctypes.windll.setupapi.SetupDiGetClassDevsW
-SetupDiGetClassDevs.argtypes = (ctypes.POINTER(GUID), ctypes.c_wchar_p, HWND, DWORD)
-SetupDiGetClassDevs.restype = ValidHandle  # HDEVINFO
-
-SetupDiGetDeviceProperty = ctypes.windll.setupapi.SetupDiGetDevicePropertyW
-SetupDiGetDeviceProperty.argtypes = (
-	HDEVINFO,  # [in]            HDEVINFO         DeviceInfoSet
-	PSP_DEVINFO_DATA,  # [in]            PSP_DEVINFO_DATA DeviceInfoData
-	ctypes.POINTER(DEVPROPKEY),  # [in]            const DEVPROPKEY *PropertyKey
-	PDWORD,  # [out]           DEVPROPTYPE      *PropertyType
-	ctypes.c_void_p,  # [out, optional] PBYTE            PropertyBuffer
-	DWORD,  # [in]            DWORD            PropertyBufferSize
-	PDWORD,  # [out, optional] PDWORD           RequiredSize
-	DWORD,  # [in]            DWORD            Flags
-)
-SetupDiGetDeviceProperty.restype = BOOL
-
-SetupDiEnumDeviceInterfaces = ctypes.windll.setupapi.SetupDiEnumDeviceInterfaces
-SetupDiEnumDeviceInterfaces.argtypes = (
-	HDEVINFO,
-	PSP_DEVINFO_DATA,
-	ctypes.POINTER(GUID),
-	DWORD,
-	PSP_DEVICE_INTERFACE_DATA,
-)
-SetupDiEnumDeviceInterfaces.restype = BOOL
-
-SetupDiGetDeviceInterfaceDetail = ctypes.windll.setupapi.SetupDiGetDeviceInterfaceDetailW
-SetupDiGetDeviceInterfaceDetail.argtypes = (
-	HDEVINFO,
-	PSP_DEVICE_INTERFACE_DATA,
-	PSP_DEVICE_INTERFACE_DETAIL_DATA,
-	DWORD,
-	PDWORD,
-	PSP_DEVINFO_DATA,
-)
-SetupDiGetDeviceInterfaceDetail.restype = BOOL
-
-SetupDiGetDeviceRegistryProperty = ctypes.windll.setupapi.SetupDiGetDeviceRegistryPropertyW
-SetupDiGetDeviceRegistryProperty.argtypes = (
-	HDEVINFO,
-	PSP_DEVINFO_DATA,
-	DWORD,
-	PDWORD,
-	ctypes.c_void_p,
-	DWORD,
-	PDWORD,
-)
-SetupDiGetDeviceRegistryProperty.restype = BOOL
-
-SetupDiEnumDeviceInfo = ctypes.windll.setupapi.SetupDiEnumDeviceInfo
-SetupDiEnumDeviceInfo.argtypes = (HDEVINFO, DWORD, PSP_DEVINFO_DATA)
-SetupDiEnumDeviceInfo.restype = BOOL
-
-CM_Get_Device_ID = ctypes.windll.cfgmgr32.CM_Get_Device_IDW
-CM_Get_Device_ID.argtypes = (DWORD, ctypes.c_wchar_p, ULONG, ULONG)
-CM_Get_Device_ID.restype = DWORD
 CR_SUCCESS = 0
 MAX_DEVICE_ID_LEN = 200